+++ src/main/java/tk/themcbros/uselessmod/lists/ModFluids.java
@@ -1,14 +1,18 @@
 package tk.themcbros.uselessmod.lists;
 
 import com.google.common.collect.Lists;
+import net.minecraft.client.renderer.RenderType;
+import net.minecraft.client.renderer.RenderTypeLookup;
 import net.minecraft.fluid.FlowingFluid;
 import net.minecraft.fluid.Fluid;
 import net.minecraft.util.ResourceLocation;
+import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.event.RegistryEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fluids.FluidAttributes;
 import net.minecraftforge.fluids.ForgeFlowingFluid;
 import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.loading.FMLEnvironment;
 import net.minecraftforge.registries.ObjectHolder;
 import tk.themcbros.uselessmod.UselessMod;
 
@@ -38,6 +42,13 @@ public class ModFluids {
 	public static void onFluidRegister(final RegistryEvent.Register<Fluid> event) {
 		FLUIDS.forEach(event.getRegistry()::register);
 		UselessMod.LOGGER.debug("Registered fluids");
+
+		if (FMLEnvironment.dist == Dist.CLIENT) {
+			RenderType translucentRenderType = RenderType.func_228645_f_();
+
+			RenderTypeLookup.setRenderLayer(USELESS_WATER, translucentRenderType);
+			RenderTypeLookup.setRenderLayer(FLOWING_USELESS_WATER, translucentRenderType);
+		}
 	}
 	
 }