+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 
 ### 1.15.1-3.0.1
 - Fluids in guis render again
+- Fixed some render type issues (still some present)
+- Fixed Large Useless Fern Loot Table
 
 ### 1.15.1-3.0.0
 - Initial 1.15.1 release